% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frank.R
\name{frank}
\alias{frank}
\title{Frank matrix of order \code{N}}
\usage{
frank(n, k = 0)
}
\arguments{
\item{n}{order of the matrix}

\item{k}{If k is 1, the elements are reflected about the anti-diagonal.}
}
\description{
Frank matrix of order \code{N}. It is upper Hessenberg with
determinant 1.
}
