% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcar.R
\name{grcar}
\alias{grcar}
\title{Create Toeplitz matrix with sensitive eigenvalues}
\usage{
grcar(n, k = NULL)
}
\arguments{
\item{n}{dimension of the square matrix}

\item{k}{number of superdiagonals of ones}
}
\value{
\code{n}-by-\code{n} Toeplitz matrix with -1 on subdiagonal, 1 on diagonal, and \code{k} superdiagionals of 1s.
}
\description{
Eigenvalues are sensitive.
}
