% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanowa.R
\name{hanowa}
\alias{hanowa}
\title{Hanowa matrix}
\usage{
hanowa(n, d = NULL)
}
\arguments{
\item{n}{order of matrix}

\item{d}{value of main diagonal}
}
\description{
Matrix whose eigenvalues lie on vertical plane in complex plane.
Returns a 2-by-2 block matrix with four \code{n/2} by \code{n/2} blocks. \code{n}
must be an even integer.

\code{[d*eye(m) -diag(1:m),}
\code{diag(1:m) d*eye(m)]}
}
