% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jordbloc.R
\name{jordbloc}
\alias{jordbloc}
\title{Create Jordan block matrix}
\usage{
jordbloc(n, lambda = 1)
}
\arguments{
\item{n}{order of matrix}

\item{lambda}{eigenvalue of Jordan block}
}
\value{
Jordan block matrix
}
\description{
Returns a \code{n}-by-\code{n} Jordan block with eigenvalue
  \code{lambda}. The default is 1.
}
