% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minij.R
\name{minij}
\alias{minij}
\title{Symmetric positive definite matrix \code{MIN(i,j)}}
\usage{
minij(n)
}
\arguments{
\item{n}{order of the matrix}
}
\description{
The \code{N}-by-\code{N} SPD matrix with \code{A[i,j]=min(i,j)}
}
