% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invol.R
\name{invol}
\alias{invol}
\title{Involutory matrix (a matrix that is its own inverse)}
\usage{
invol(n)
}
\arguments{
\item{n}{order of matrix}
}
\value{
Involutory matrix (a matrix that is its own inverse).
}
\description{
a \code{n}-by\code{n} involutory matrix and ill-conditioned.
It is a diagonally scaled version of a Hilbert matrix.
}
